+++ src/main/java/com/authlete/jaxrs/spi/TokenRequestHandlerSpiAdapter.java
@@ -54,4 +54,11 @@ public class TokenRequestHandlerSpiAdapter implements TokenRequestHandlerSpi
     {
         return null;
     }
+
+
+    @Override
+    public Response jwtBearer(TokenResponse tokenResponse)
+    {
+        return null;
+    }
 }